requirements.txt
fastapi==0.109.0
supabase==2.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
